requirements.txt
google-generativeai>=0.8.3
pypdf
pandas
openpyxl
fastapi
uvicorn
streamlit
python-dotenv
python-multipart
groq
